Move ObsSummaryRow and the obs summary column defs to separate files
